Separació environment producció i desenvolupament
