+++ pom.xml
@@ -15,6 +15,24 @@
 
     <description>TFG project for Spring Boot</description>
     
+    <profiles>
+    	<profile>
+    		<id>dev</id>
+    		<properties>
+    			<spring.profiles.active>dev</spring.profiles.active>
+    		</properties>
+    		<activation>
+    			<activeByDefault>true</activeByDefault>
+    		</activation>
+    	</profile>
+    	<profile>
+    		<id>prod</id>
+    		<properties>
+    			<spring.profiles.active>prod</spring.profiles.active>
+    		</properties>
+    	</profile>	
+    </profiles>
+    
 	<repositories>
 		<repository>
 		    <id>jitpack.io</id>
