+++ src/main/java/com/webapp/gessi/GESSIApplication.java
@@ -4,7 +4,6 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.web.bind.annotation.CrossOrigin;
 
 @SpringBootApplication
 public class GESSIApplication extends SpringBootServletInitializer {
@@ -15,6 +14,9 @@ public class GESSIApplication extends SpringBootServletInitializer {
     }
 
     public static void main(String[] args) {
+    	for (String arg: args) {
+    		System.out.println(arg); 		
+    	}
         SpringApplication.run(GESSIApplication.class, args);
     }
