+++ src/main/resources/application-dev.properties
@@ -1,0 +1,3 @@
+spring.datasource.url=jdbc:derby://localhost:1033/DOCKERDB;
+spring.datasource.username=DOCKERDB
+spring.datasource.password=DOCKERDB
