+++ src/main/resources/application-prod.properties
@@ -1,0 +1,3 @@
+spring.datasource.url=jdbc:derby://gessi-derby:1527/DOCKERDB;
+spring.datasource.username=GESSI-SLR
+spring.datasource.password=dbgessi1.2.0&
