+++ src/main/resources/application.properties
@@ -1,9 +1,5 @@
-#spring.datasource.url=jdbc:derby://gessi-derby:1527/DOCKERDB;
-spring.datasource.url=jdbc:derby://localhost:1033/DOCKERDB;
-spring.datasource.username=DOCKERDB
-spring.datasource.password=DOCKERDB
+spring.profiles.active = @spring.profiles.active@
 spring.mvc.view.prefix=src/main/resources/static
 spring.mvc.view.suffix=.html
 spring.servlet.multipart.max-file-size=100MB
-#spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.web.servlet.error.ErrorMvcAutoConfiguration
 server.error.path=/api/error
